Update GroupGridView.php
Fix: Providing right context for the evaluation of the column value.
Although the code is well copied from renderDataCell in framework/zii/widgets/grid/CGridColumn.php#135, the data variable is not in the same scope. 

First line in original code is missing: $data=$this->grid->dataProvider->data[$row];
I added a similar one.